pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "test_availability_data"
version = "0.1.1"
description = "Script for testing availability of data of the Marine data store via the toolbox"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13"
authors = [
  {name = "Guillaume Koenig", email = "gkoenig@mercator-ocean.fr"}
]

[tool.poetry.dependencies]
copernicusmarine = "==2.3.0"
pandas = "==2.3.3"
numpy = "==2.4.2"
sqlalchemy = "==2.0.46"
psycopg2-binary = "==2.9.10"
zarr = "==3.1.5"
tabulate = "==0.9"
distro = "==1.9.0"
xarray = "==2026.1.0"
python-dotenv = "^1.2.2"


[tool.poetry.group.dev.dependencies]
mkdocs = "==1.6.1"
responses = ">=0.25.0"
pytest = "^9.0.3"

[tool.poetry.group.deploy.dependencies]
requests = "^2.32.5"
json5 = "^0.13.0"

